LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/translate-getComputedStyle-expected.txt


FAIL computed style for translate assert_equals: expected "30% 40% 50px" but got "0px 0px 50px"
